# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Concord - NLP-powered task management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /cycle:
    get:
      operationId: getCycleInfo
      tags: [profile]
      summary: Get current menstrual cycle phase and efficiency
      responses:
        "200":
          description: Cycle phase information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleInfo"
        "401":
          description: Unauthorized
  /tasks/parse:
    post:
      operationId: parseTask
      tags: [tasks]
      summary: Parse natural language input into a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseInput"
      responses:
        "200":
          description: Parsed task preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedTask"
        "400":
          description: Bad request
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
  /tasks/{id}/status:
    patch:
      operationId: updateTaskStatus
      tags: [tasks]
      summary: Update task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [userId, profileCompleted]
      properties:
        userId:
          type: string
        displayName:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [female, male, other, prefer_not_to_say, null]
        lastPeriodDate:
          type: ["string", "null"]
          format: date-time
        cycleLength:
          type: integer
        lutealPhaseLength:
          type: integer
        lastPeriodUpdatedAt:
          type: ["string", "null"]
          format: date-time
        profileCompleted:
          type: boolean
        weekendAvailable:
          type: boolean
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        gender:
          type: string
          enum: [female, male, other, prefer_not_to_say]
        lastPeriodDate:
          type: ["string", "null"]
          format: date-time
        cycleLength:
          type: integer
        lutealPhaseLength:
          type: integer
        weekendAvailable:
          type: boolean
    CycleInfo:
      type: object
      required: [phase, dayInCycle, efficiency, hasData]
      properties:
        hasData:
          type: boolean
        phase:
          type: string
          enum: [menstrual, follicular, ovulatory, luteal, unknown]
        dayInCycle:
          type: ["integer", "null"]
        efficiency:
          type: integer
          description: Percentage efficiency 0-100
        phaseLabel:
          type: string
        phaseDescription:
          type: string
        shouldReschedule:
          type: boolean
          description: True during luteal and menstrual phases
        needsPeriodUpdate:
          type: boolean
          description: True if last period date was updated more than cycleLength days ago
    ParseInput:
      type: object
      required: [rawInput]
      properties:
        rawInput:
          type: string
          description: Natural language task description
    ParsedTask:
      type: object
      required: [title, rawInput]
      properties:
        rawInput:
          type: string
        title:
          type: string
        summary:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        deadline:
          type: ["string", "null"]
          format: date-time
        keywords:
          type: array
          items:
            type: string
    Task:
      type: object
      required: [id, title, rawInput, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        title:
          type: string
        rawInput:
          type: string
        summary:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        deadline:
          type: ["string", "null"]
          format: date-time
        keywords:
          type: array
          items:
            type: string
        scheduledDate:
          type: ["string", "null"]
        duration:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title, rawInput]
      properties:
        rawInput:
          type: string
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        deadline:
          type: ["string", "null"]
          format: date-time
        keywords:
          type: array
          items:
            type: string
        scheduledDate:
          type: string
        duration:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        deadline:
          type: ["string", "null"]
          format: date-time
        keywords:
          type: array
          items:
            type: string
        scheduledDate:
          type: string
        duration:
          type: string
    TaskStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed]
...